Refine brake test example
